Rework negation function
The negate function still sucks, but it kind of works. But it has
limitations, e.g. you can't change the sign before entering anything.
It seems to me to be an overall design issue. Implementing this
feature would be a lot easier if I keept the results of calculation
as currentTerm.