Finish layout and fix minor bugs
I added new fonts and made final adjustments to the layout and style
of the calculator.

The maximum number of digits that can be entered had to be limited
by 14. Also, CE button didn't need to remove the result of the
calculation after it was finished - now it does nothing if
finishedEvaluation is true.